requirements.txt
fastapi
supabase
google-generativeai
python-dotenv
pydantic
pydantic-settings
postgrest
